--- src/screens/proposal_details/components/detail/index.test.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import { mount } from 'enzyme';
-import { BaseWrapper } from '@tests/utils/base_wrapper';
-import { lightTheme } from '@styles';
-import Deposit from '.';
-
-describe('Deposit', () => {
-  it('correctly renders component', () => {
-    const wrapper = mount(
-      BaseWrapper({
-        component: <Deposit />,
-        theme: lightTheme,
-      }),
-    );
-    expect(wrapper).not.toBeNull();
-    expect(wrapper.find(Deposit)).toHaveLength(1);
-  });
-});
