+++ src/screens/proposal_details/components/index.ts
@@ -1,9 +1,7 @@
 import Deposit from './deposit';
-import Detail from './detail';
 import Info from './info';
 
 export {
   Deposit,
   Info,
-  Detail,
 };
