+++ src/screens/proposal_details/components/info/components/index.tsx
@@ -1,0 +1,5 @@
+import Proposer from './proposer';
+
+export {
+  Proposer,
+};
