+++ src/screens/proposal_details/components/info/components/proposer/styles.tsx
@@ -1,0 +1,18 @@
+import {
+  makeStyles, createStyles,
+} from '@material-ui/styles';
+
+export const useGetStyles = () => {
+  const useStyles = makeStyles((theme: any) => createStyles({
+    root: {
+      display: 'flex',
+      '& div': {
+        color: theme?.palette?.custom?.fonts?.fontFive,
+      },
+    },
+  }));
+
+  return {
+    classes: useStyles(),
+  };
+};
