+++ src/screens/proposal_details/components/info/components/proposer/types.ts
@@ -1,0 +1,6 @@
+export interface ProposerProps {
+  id: string;
+  imageUrl: string;
+  display: string;
+  alt?: string;
+}
