+++ src/screens/proposal_details/components/info/index.test.tsx
@@ -2,17 +2,17 @@ import React from 'react';
 import { mount } from 'enzyme';
 import { BaseWrapper } from '@tests/utils/base_wrapper';
 import { lightTheme } from '@styles';
-import Deposit from '.';
+import Info from '.';
 
 describe('Deposit', () => {
   it('correctly renders component', () => {
     const wrapper = mount(
       BaseWrapper({
-        component: <Deposit />,
+        component: <Info />,
         theme: lightTheme,
       }),
     );
     expect(wrapper).not.toBeNull();
-    expect(wrapper.find(Deposit)).toHaveLength(1);
+    expect(wrapper.find(Info)).toHaveLength(1);
   });
 });
