+++ src/screens/proposal_details/components/info/styles.tsx
@@ -9,30 +9,45 @@ export const useGetStyles = () => {
       borderRadius: '6px',
       background: theme?.palette?.background?.paper,
       '&.big-dipper': {
-        '& h1': {
-          color: theme?.palette?.custom?.fonts?.fontOne,
-        },
-        '& .cell': {
-          color: theme?.palette?.custom?.fonts?.fontTwo,
-        },
+        // '& h1': {
+        //   color: theme?.palette?.custom?.fonts?.fontOne,
+        // },
+
         '& .tableRow': {
           '&:nth-of-type(odd)': {
             background: theme?.palette?.type === 'light' ? '#F8F8F8' : '#19191D',
           },
         },
       },
-      '&.proposalDepositMobile': {
+      '&.proposalInfoMobile': {
         padding: theme?.palette?.custom?.margins?.small,
+        paddingTop: '0',
       },
-      '&.proposalDepositDesktop': {
+      '&.proposalInfoDesktop': {
         display: 'none',
       },
       [getMinMediaQuery(theme?.breakpoints?.values?.tablet)]: {
-        '&.proposalDepositDesktop': {
+        '&.proposalInfoDesktop': {
+          background: theme?.palette?.background?.paper,
           padding: theme?.palette?.custom?.margins?.large,
           display: 'block',
+          '& .title': {
+            color: theme?.palette?.custom?.fonts?.fontOne,
+          },
+          '& .cell, & .label, & .time': {
+            color: theme?.palette?.custom?.fonts?.fontTwo,
+          },
+          '& .table': {
+            borderBottomColor: theme?.palette?.divider,
+          },
+          '& .days': {
+            color: theme?.palette?.type === 'light' ? '#FD7522' : '#FD7522B3',
+          },
+          '& .button': {
+            color: theme?.palette?.type === 'light' ? '#FFFFFF' : theme?.palette?.custom?.fonts?.fontOne,
+          },
         },
-        '&.proposalDepositMobile': {
+        '&.proposalInfoMobile': {
           display: 'none',
         },
       },
