+++ src/screens/proposal_details/components/info/utils.tsx
@@ -1,95 +1,46 @@
 import React from 'react';
-import { AvatarDisplay } from 'big-dipper-default-ui';
+// import { AvatarDisplay } from 'big-dipper-default-ui';
+import { Proposer } from './components';
 
-export const depositData = [
-  {
-    depositor: <AvatarDisplay
+const Proposer1 = () => {
+  return (
+    <Proposer
       imageUrl="https://s3.amazonaws.com/keybase_processed_uploads/f5b0771af36b2e3d6a196a29751e1f05_360_360.jpeg"
       alt="avatar image"
-      title="Forbole"
-    />,
-    amount: '10.8 ATOM',
-    time: '09 Jan 2020, 10:00:19',
+      id="1"
+      display="Forbole"
+    />
+  );
+};
+
+export const dummyData: any = {
+  id: '01',
+  proposer: <Proposer1 />,
+  title: 'Are Validators Charging 0% Commission Harmful to the Success of the Cosmos Hub?',
+  description: 'Lunie Mobile now available in app stores near you! 📲 iOS - bit.ly 📲 Android - bit.ly ❣️This proposal was sent from the Lunie iOS app 😏Lunie Mobile now available in app stores near you! 📲 iOS - bit.ly 📲 Android - bit.ly ❣️This proposal was sent from the Lunie iOS app 😏Lunie Mobile now available in app stores near you! 📲 iOS - bit.ly 📲 Android - bit.ly ❣️This proposal was sent from the Lunie iOS app 😏Lunie Mobile now available in app stores near you! 📲 iOS - bit.ly 📲 Android - bit.ly ❣️This proposal was sent from the Lunie iOS app 😏',
+  submittedTime: '12 Dec 2019, 16:22:11 UTC',
+  depositEndTime: '26 Dec 2019, 16:22:11 UTC',
+  votingStartTime: '26 Dec 2019, 16:22:11 UTC',
+  votingEndTime: '10 Jan 2020, 16:22:11 UTC',
+  status: {
+    current: false,
+    display: 'Passed',
   },
-  {
-    depositor: <AvatarDisplay
-      imageUrl="https://s3.amazonaws.com/keybase_processed_uploads/f5b0771af36b2e3d6a196a29751e1f05_360_360.jpeg"
-      alt="avatar image"
-      title="Forbole"
-    />,
-    amount: '10.9 ATOM',
-    time: '09 Jan 2020, 10:00:20',
-  },
-  {
-    depositor: <AvatarDisplay
-      imageUrl="https://s3.amazonaws.com/keybase_processed_uploads/f5b0771af36b2e3d6a196a29751e1f05_360_360.jpeg"
-      alt="avatar image"
-      title="Forbole"
-    />,
-    amount: '5.8 ATOM',
-    time: '09 Jan 2020, 12:00:19',
-  },
-  {
-    depositor: <AvatarDisplay
-      imageUrl="https://s3.amazonaws.com/keybase_processed_uploads/f5b0771af36b2e3d6a196a29751e1f05_360_360.jpeg"
-      alt="avatar image"
-      title="Forbole"
-    />,
-    amount: '16.8 ATOM',
-    time: '09 Jan 2020, 10:30:19',
-  },
-  {
-    depositor: <AvatarDisplay
-      imageUrl="https://s3.amazonaws.com/keybase_processed_uploads/f5b0771af36b2e3d6a196a29751e1f05_360_360.jpeg"
-      alt="avatar image"
-      title="Forbole"
-    />,
-    amount: '6.8 ATOM',
-    time: '10 Jan 2020, 10:00:19',
-  },
-  {
-    depositor: <AvatarDisplay
-      imageUrl="https://s3.amazonaws.com/keybase_processed_uploads/f5b0771af36b2e3d6a196a29751e1f05_360_360.jpeg"
-      alt="avatar image"
-      title="Forbole"
-    />,
-    amount: '10.8 ATOM',
-    time: '09 Jan 2020, 10:00:19',
-  },
-  {
-    depositor: <AvatarDisplay
-      imageUrl="https://s3.amazonaws.com/keybase_processed_uploads/f5b0771af36b2e3d6a196a29751e1f05_360_360.jpeg"
-      alt="avatar image"
-      title="Forbole"
-    />,
-    amount: '10.9 ATOM',
-    time: '09 Jan 2020, 10:00:20',
-  },
-  {
-    depositor: <AvatarDisplay
-      imageUrl="https://s3.amazonaws.com/keybase_processed_uploads/f5b0771af36b2e3d6a196a29751e1f05_360_360.jpeg"
-      alt="avatar image"
-      title="Forbole"
-    />,
-    amount: '5.8 ATOM',
-    time: '09 Jan 2020, 12:00:19',
-  },
-  {
-    depositor: <AvatarDisplay
-      imageUrl="https://s3.amazonaws.com/keybase_processed_uploads/f5b0771af36b2e3d6a196a29751e1f05_360_360.jpeg"
-      alt="avatar image"
-      title="Forbole"
-    />,
-    amount: '16.8 ATOM',
-    time: '09 Jan 2020, 10:30:19',
-  },
-  {
-    depositor: <AvatarDisplay
-      imageUrl="https://s3.amazonaws.com/keybase_processed_uploads/f5b0771af36b2e3d6a196a29751e1f05_360_360.jpeg"
-      alt="avatar image"
-      title="Forbole"
-    />,
-    amount: '6.8 ATOM',
-    time: '10 Jan 2020, 10:00:19',
-  },
-];
+  // duration: '(In 12 days)',
+  type: 'Text proposal',
+  deposit: '520 ATOM',
+};
+
+export const Labels: any = {
+  id: 'Proposal ID',
+  proposer: 'Proposer',
+  type: 'Type',
+  title: 'Title',
+  description: 'Description',
+  deposit: 'Deposit',
+  status: 'Status',
+  submittedTime: 'Submitted Time',
+  depositEndTime: 'Deposit End Time',
+  votingStartTime: 'Voting Start Time',
+  votingEndTime: 'Voting End Time',
+};
